requirements.txt
cffi==1.17.1
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.2.1
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
qrcode==8.1
six==1.17.0
